requirements.txt
# run
pyyaml

# dev
pytest
codecov
coverage>=5
pytest-cov
pylint
docmd
black
pre-commit
